+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ This is the offical release of vpinfe v1.1.38.
 
 ### What's New
 - updated the theme doc to account for the new bg and dmd high DPI overrides
+- added "Score" section to .info
 
 ### Fixes
 None